bug fix: lombok problem

+++ src/main/java/com/example/parking_management_system_api/repositories/VehicleRepository.java
@@ -3,5 +3,5 @@ package com.example.parking_management_system_api.repositories;
 import com.example.parking_management_system_api.entities.Vehicle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleRepository extends JpaRepository<Vehicle, Integer> {
+public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
 }
